tests/wpt/tests/html/semantics/grouping-content/the-figcaption-element/grouping-figcaption.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>the figcaption element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-figcaption-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        "use strict";

        // check that prototype matches spec's DOM interface
        test(function () {
            var testElement = document.createElement("figcaption");
            assert_equals(Object.getPrototypeOf(testElement), HTMLElement.prototype, "HTMLElement.prototype should be used for figcaption");
        }, "The prototype for figcaption is HTMLElement.prototype");

    </script>
</head>
<body>
    <h1>Description</h1>
    <p>This test validates the figcaption element.</p>

    <div id="log"></div>

</body>
</html>
